f64 -> f32

+++ src/math/ray.rs
@@ -1,12 +1,12 @@
 use crate::math::vec3::Vector3;
 
 pub struct Ray {
-    pub origin: Vector3<f64>,
-    pub direction: Vector3<f64>,
+    pub origin: Vector3<f32>,
+    pub direction: Vector3<f32>,
 }
 
 impl Ray {
-    pub fn new(origin: Vector3<f64>, direction: Vector3<f64>) -> Self {
+    pub fn new(origin: Vector3<f32>, direction: Vector3<f32>) -> Self {
         return Ray {origin, direction} ;
     }
 }
