+++ src/renderer/camera.rs
@@ -2,15 +2,15 @@ use crate::math::vec3::Vector3;
 use crate::renderables::renderable::Renderable;
 
 pub struct Camera {
-    position: Vector3<f64>,
-    direction: Vector3<f64>,
+    position: Vector3<f32>,
+    direction: Vector3<f32>,
     fov: f32,
     z_near: f32,
     z_far: f32,
 }
 
 impl Camera {
-    pub fn new(fov: f32, z_near: f32, z_far: f32, position: Vector3<f64>, direction: Vector3<f64>) -> Self {
+    pub fn new(fov: f32, z_near: f32, z_far: f32, position: Vector3<f32>, direction: Vector3<f32>) -> Self {
         Camera {
             fov,
             z_far,
@@ -20,7 +20,7 @@ impl Camera {
         }
     }
 
-    pub fn position(&self) -> &Vector3<f64> {
+    pub fn position(&self) -> &Vector3<f32> {
         &self.position
     }
